refactor(floating-btn): Extract accent color and theme management logic
- Split accent color logic and theme management logic from the floating button file into a separate file for improved readability and maintainability

Changes summary:
- Extracted accent color and theme management logic from the floating button file:
  - Improved code organization and readability by separating concerns into distinct files.
  - Enhanced maintainability by making it easier to update and modify accent color and theme management logic independently.
