+++ src/js/app/layouts/settings.js
@@ -1,0 +1,28 @@
+// // import { renderColorsTab, handleColorsListeners } from './mainColors.js'
+// import { renderFontsTab, handleFontsListeners } from './mainFonts.js'
+// import { renderAssetsTab, handleAssetsListeners } from './mainAssets.js'
+
+// const renderSettings = `
+//     <header class="mb-5">
+//       <h2 class="mt-5 text-center font-medium gpth-settings__title"><span class="font-semibold">GPThemes</span> Customization</h2>
+//       <button class="text-token-text-tertiary hover:text-token-text-primary absolute top-4 right-4" id="gpth-settings-close">
+//         <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M6.34315 6.34338L17.6569 17.6571M17.6569 6.34338L6.34315 17.6571" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"></path></svg>
+//       </button>
+//     </header>
+//     <main>
+//       <div class="tabs">
+//         <div class="tab-buttons flex items-center rounded-full p-1 font-semibold mb-10">
+//           <button class="tab-button py-2 px-4 focus:outline-none text-center rounded-full active">Color</button>
+//           <button class="tab-button py-2 px-4 focus:outline-none text-center rounded-full">Font</button>
+//           <button class="tab-button py-2 px-4 focus:outline-none text-center rounded-full">Assets</button>
+//         </div>
+//         <div class="tab-content">
+//           <div class="tab-pane active" id="tab-colors">${renderColorsTab}</div>
+//           <div class="tab-pane hidden" id="tab-fonts">${renderFontsTab}</div>
+//           <div class="tab-pane hidden" id="tab-assets">${renderAssetsTab}</div>
+//         </div>
+//       </div>
+//     </main>
+//   `
+
+// export { renderSettings }
