+++ src/js/content.js
@@ -1,7 +1,11 @@
+import { init as initThemes } from './app/themeManager'
 import { init as initFloating } from './app/floatingBtn'
+import { init as initColors } from './app/mainColors'
 import { init as initFonts } from './app/mainFonts'
 import { init as initAssets } from './app/mainAssets'
 
+initThemes()
 initFloating()
+initColors()
 initFonts()
 initAssets()
